+++ providers/api.rb
@@ -5,7 +5,7 @@ def whyrun_supported?
 end
 
 def health_check
-  ruby_block 'wait for Nexus Rest API endpoint to come up' do
+  ruby_block 'wait for Nexus Rest API endpoint to respond' do
     block do
       i = 0
       wait = new_resource.wait
@@ -54,8 +54,6 @@ EOF
     only_if { new_resource.script_source.nil? }
   end
 
-  install_curl
-
   execute "upload script #{new_resource.script_name}" do
     command "curl -v #{fail_silently} -X POST -u #{new_resource.username}:#{new_resource.password}" \
         " --header \"Content-Type: application/json\" '#{new_resource.endpoint}' -d @#{new_resource.script_name}.json"
@@ -87,8 +85,6 @@ end
 def run_script
   create_script unless new_resource.script_source.nil? && new_resource.content.nil?
 
-  install_curl
-
   execute "run script #{new_resource.script_name}" do
     command "curl -v #{fail_silently} -X POST -u #{new_resource.username}:#{new_resource.password}" \
         " --header \"Content-Type: text/plain\" '#{new_resource.endpoint}/#{new_resource.script_name}/run' #{args}"
@@ -99,8 +95,6 @@ def run_script
 end
 
 def delete_script
-  install_curl
-
   execute "delete script #{new_resource.script_name}" do
     command "curl -v -X DELETE -u #{new_resource.username}:#{new_resource.password}" \
       " '#{new_resource.endpoint}/#{new_resource.script_name}'"
@@ -124,8 +118,6 @@ def list_scripts
     mode '0755'
     action :create
   end
-
-  install_curl
 
   execute "write #{list_file}" do
     command "curl -v -X GET -u #{new_resource.username}:#{new_resource.password} '#{new_resource.endpoint}'" \
@@ -155,6 +147,7 @@ end
 
 action :run do
   converge_by('run script') do
+    install_curl
     health_check
     run_script
   end
@@ -162,6 +155,7 @@ end
 
 action :create do
   converge_by('create script') do
+    install_curl
     health_check
     create_script
   end
@@ -169,6 +163,7 @@ end
 
 action :delete do
   converge_by('delete script') do
+    install_curl
     health_check
     delete_script
   end
@@ -176,6 +171,7 @@ end
 
 action :list do
   converge_by('list scripts') do
+    install_curl
     health_check
     list_scripts
   end